Fix unnecesary timeout on various jobs cancel tests
Use `wait_until_completed()` to first assert that the project_update has
finished.  Otherwise, we wait for the provided `timeout` (4 mins in this
case) in the event the `project_update` completes, but is not `status ==
canceled`.  This resolves issues where the project_update completes
('failed' in this case), but the test continues to wait for status ==
'canceled' for 4 minutes.