Add jenkins-jnlp-agent
This is an alternative to the strategy used in `jenkins-swarm-agent`. Now that the Jenkins Kubernetes Plugin supports [multiple containers](https://github.com/jenkinsci/kubernetes-plugin/pull/67), it makes this a feasible (probably preferable) approach because the agents come online, do their work, and then die. This means a fresh environment on every build.