ensure task manager fail and cancel chain behavior
* The task manager has been changed to cancel and fail update_on_launch
tasks together. For example, launch a JT with proj scm_update_on_launch
and inventory source update_on_launch. Cancel anyone one of the spawned
jobs and all jobs in that "unit" should be canceled. Same with failure,
if a single job from that "unit" fails, all jobs in that "unit" should
fail.