In network partition test, when preparing job template during partition, run project update on local partition (#1473)
After test_network_partition creates a network partition, it attempts to launch a job that is assigned to run against the other partition. The goal is to ensure that the job eventually reaches a completed state. The test is currently failing when preparing a project to be used by the job template. I suspect this is because the project update is attempting to run on an instance in the other partition. Since I am interested in testing the behavior of a job that is launched and not the project update, I'm updating the test so that the project update will run on the current partition.