Update tests to permit error status as well as failed
A long running isolated job that loses touch with a tower HA cluster, because all controller nodes are down, is now reaped as being in an error status rather than a failure status as was previously the case. We don't care about the specific failure mode, only that the job is recognised as being "lost" in a timely manner.