remove flakey status=pending websocket checks
this is more test flake;  *sometimes* these tests fail because we never
get a `pending` event.  This feels like a scenario where the pending ->
waiting state happens so quickly that we sometimes just don't get the
message.  This _could_ just come down the asynchronous nature of our
websocket connection and message handling.  In any event, it's a silly
thing to gate this complicated integration test on, since it's not the
point anyways (RBAC of the messages is).

I'm able to make these tests fail intermittently, but only if I run the
full suite in a loop for 20+ minutes.

In the end, if we get a `waiting` *and* `running` *and* `successful`
message, it's pretty clear that this is working as intended.