do not wait for job events to be processed
* The helper method wait_until_completed() waits for a job's status
state to be a completed one (i.e. successful, failed, error) AND waits
for all events for the job to be processed. We don't want to wait for
all events to be processed on our job. We hard kill the openshift
container and this can result in events not being processed.