Test_Tower_Upgrade: Make the job cluster compatible
Current Test_Tower_Upgrade can only work for standalone install. A
standalone or a cluster install should be treated equally, in the sense
that from an installer perspective it is exactly the same process.

This commits allow one to specify the playbook to run (ie.
deploy-tower.yml vs deploy-tower-cluster.yml) as the
TOWER_INSTALL_PLAYBOOK environment variable.