isolated_group_firewall: Ensure proper IP is used for SSH rules
With a recent change in the inventory used (ie. having different
inventory_hostname and ansible_host), ansible_default_ipv4 defaults to
the ec2 instance *private* ip. Which leads to locking isolated nodes
from the rest of the planet - including controller nodes.

In order to avoid that, we ensure the public IP is used for the SSH
rules. The use of ipyfi here is due to the fact that no gathered fact
had the proper value.