Update and add tests regarding the results of having canceled nodes in a WF
These tests were catching a regression but it was difficult to understand what
was happening.

If a WF node is canceled, it is treated as a failure. If it has an always or
failure node that runs successfully, it is considered "handeled" and so the WFJ
is successful. If it is not "handeled" by a always or failure node, the WFJ is
marked "failed."