Properly skip_if_enabled fixture on OpenShift
Some of our tests are now conditioned by a skip_if_fips_enabled fixture.

What this fixture does is that it gathers facts on the node listed in
`--ansible-host-pattern` and look at `ansible_fips` fact on those nodes.

The `--ansible-host-pattern` specified in the testing command line is
not part of the Inventory. Hence the issue:

```
AttributeError: type HostManager has no attribute
'ansible-tower-web-svc-tower-qe.openshift.ansible.eng.rdu2.redhat.com'
```

In an OpenShift scenario, before the introduction of
`skip_if_fips_enabled` no test was relying on any fixture of
`pytest-ansible` so to bypass this limitation we can add an harmless
hostname.