Counting only revelant events
Some tests count the number of events and make an assumption on the
number of events that should be returned.

Starting with 2.8.0, ansible returns more events than before (it started
returning `runner_on_start` events). In order to still work on both
pre-2.8 and 2.8+ we specify the type of event we want or do not want in
order to make the assertions.