make dispatcher killing a bit less flaky
there are a few changes here, but the main issue is that we need to wait
a little longer before killing the dispatcher if we want to test this in
the way we're doing it here. What's happening here:

1. We launch a Job.
2. A dispatcher worker grabs it, and places it into running state.
3. Before the actual playbook runs, we generate a sync Project Update:
   https://github.com/ansible/awx/blob/devel/awx/main/tasks.py#L1378
   ...and it begins running.
4. We immediately stop the dispatcher. The newly created Project Update
   and the job are **both** reaped and marked as failed because they're
   **both** technically in `status='running'` at this point.
5. Because the Project Update is marked as failed, its related project
   is _also_ set to status=failed.
6. Later on in this test, we try to launch the JT again, and it is
   unhappy that the Project status is failed (due to this recent change:
   ansible/awx@b790b50)

If we instead wait for at least one job event to show up, we know that
the playbook itself has started running, and we won't run into this race
where we prematurely put the Project into a failed state.