Tests: Retrieving the proper ansible/ansible
Currently there is a fixture that retrieves the `ansible/ansible`
project, but it retrieves `devel` no matter which version of ansible is
installed.

`devel` is an always mooving target, this behavior might break our tests
when the ansible version  installed becomes uncompatible with `devel`
this patch ensures the propre version of `ansible/ansible` is retrieved.