Migrate from ec2 to ec2_instance module
In the create_ec2 and terminate_ec2 roles, the use of `ec2` module has
been replaced with the use of `ec2_instance` module.

`ec2` module is considered legacy and `ec2_instance` is the "modern" way
of using it. The core reason why this change is needed in our case is
that by using `ec2_instance` module we can rely on the auto-assign ipv6
IP which did not work with `ec2` module.