Create Python 3 virtualenvs on Jenkins home
Bash shebang's length is limited to 128 characters and we can easily go
over that on our Jenkins jobs since we are using the working directory.
Update the Python 3 virtualenv creation to be done on Jenkins' home
instead of the job's working directory.

Update the `setup_python3_env` to not install any requirements, leave
this to be done outside the function.