Ubuntu: Fix python race condition
Initially we were waiting for 90sec before doing anything after
create_ec2 created the VMS. Meanwhile python was being installed via
cloud-init.

Now that we don't wait for those 90sec, but proceed as soon as OpenSSH
is ready, we end up trying to access the box before python is installed
in it which leads to the ansible_python_interpreter not being found.

This commit makes it explicit the way we deal with python on Ubuntu and
removes the user-data mechanism.