Verify: Do not verify script content
Due to a change in tower the content of the script can evolve from a
release to another. Fox example from 3.4 to 3.5 script moved to python3,
requiring to change `print x` to `print(x)` leading to the verify to be
broken.

As long as the metadata are accurate we will consider the inventory
script to be accurate.