Parametrize tests to also test isolated nodes
Add a fixture that will return first the tower instance group and then the
isolated instance group. We can then assign job templates these
instance groups to run on, effectively making the test run once on
regular instances, a second time on isolated ones. It will skip the
second one if not on a cluster.