Test that urls are properly sanitized in job_env
Acheive this by creating a custom credential with two different types of
urls that contain secrets. Confirm that the secrets are santized.

Also update to use v2 job templates and change order in which
instance_groups fixture is consumed to save time so if the isolated_node
parameterization is going to skip, it does it as soon as possible.