test_inventory_update: Fix ansible version cmp
In the test_inventory_update test file, the ansible version comparison
is erroneous leading to failure when the test suite is run.

To ensure test condition are met in 2.8.0, the
`ansible_version_cmp('2.8.0') >= 1` is used. But when tower is running
with `2.8.0` this method returns 0.

```
>>> towerkit.tower.version_cmp('2.8.0', '2.8.0')
0
```

As used in most places in `ansible/tower-qa` this pattern has been
replaced with `ansible_version_cmp('2.8.0') < 0` to make proper
determination of previous ansible version than 2.8.0.