Refactor data verfication script to be tests
Uses pytest and existing fixtures to deal with authentication and API access.

The previous incarnation of this produced unreadable output and was very painful to interact with, not allowing the
user to run only parts that failed, and debugging failures was very difficult.

Now we can run only failing tests, as well as use common pytest tools like `--pdb` to drop into a pdb session
on assertion failure.

These data verification tests can be run as many times as you like on a loaded tower.

Next step is to port the data loading to a similar framework and make it indempotent.