Yolo: Re-enable live stream for install and test
When we moved to the script based approach rather than the Job based
approach that was in place before we lost the live stream feature.

This was due because, we - via ansible - were calling a script to be run
from the test-runner node. Since ansible can't stream results back we
couldn't do anything but wait for the full stream to display it.

Now, rather than calling ansible to call the script we can call it
directly via ssh, this allow us to have the live stream back.