LP Testing: Reorganize input parameter
Update the LP Testing pipeline to it takes a y-stream product version
rather than a z-stream.

This is done so it can easily be plugged to the LP Testing event being
sent by the interop team. The pipeline will be in charge of deducing the
z-stream version just like verification.groovy.