job_template_with_random_attributes: Always assign project
This test is sometimes failing. This is due that if the job_type is
assigned to 'scan' no project is then assigned[1]

This leads to the following error:

```
awxkit.exceptions.BadRequest: {'project': ['Job Templates must have a project assigned.']}
```

This does not happen if job_type isn't scan. So we ensure this test
passes 100% of the time.

[1]
https://github.com/ansible/awx/blob/devel/awxkit/awxkit/api/pages/job_templates.py#L123-L124