terminate_ec2: Remove ambiguity on when condition
Before this commit, this role was called as a dependency of `create_ec2`
role. This lead to both value `delete_on_start` *or* `reap_instances` to
have an impact on the teardown process.

So for example if one selected `delete_on_start` True but
`reap_instances` False, Vms would have been deleted either way.

This new approach decouple both actions.