Address timeout flake in restart test
We have been getting failures where "wait_until_status" was timing out afte "4.76837158203125e-07 seconds.",
which is very short! This seems to be measuring from job start instead of from beginning of call.

Now, wait until it is completed and then assert the status. The wait_until_completed call starts the clock
from the time at which it was called.
For reference, see:
 https://github.com/ansible/awx/blob/b2b475d1a6a9c9788e10cc5aff8e92589fd56840/awxkit/awxkit/api/mixins/has_status.py#L34