Allow empty results for API and CLI jUnit report
YOLO uses the same test expression for both API and CLI tests. It can
happen that the test expression does not match tests on both API and CLI
and therefore one of those will get empty results. This will allow empty
results when collecting the API and CLI jUnit reports.