Start on container groups fixtures
Goal of this is to get MVP set up to have our gke cluster setup to be used as a
container group in tower under test.

Seems to be MOSTLY working but getting 401 authentication error when I run the job --
I'm guessing that means the cert is good but the token is bad/wrong one or I'm missing something.

decode the token or it does not work

Fix image spec for container group