Extend parametrized job_template to also run on CG
Previosly we enabled running all these tests on isolated instance groups
to test all the features that are not "isolated specific" work also on
isolated nodes. Extending this approach to container groups, so we can
confirm the image we ship and method we use works just as well as
regular tower instance.