Skip awx collection tests in a clustered deployment
In a clustered deployment, the node we want to install the awx
collection on doesnt have the ssh keys to access the tower repo. This
means when we deploy SLOWYO with tower, the git clone to install awx
collection just hangs waiting for the username and password to access
github. Instead lets just skip these tests on clustered deployments