Improve error message for code obfuscation test
Previously, it was difficult to know what happened if we did not get expected exit code
Now we get more details to allow us to know the nature of the change that caused the bug to fail.

Also update to use more reliable version of ansible_runner that is easier to control what hosts it runs against.